src/api/motogp_client.py
"""
MotoGP API Client
Handles communication with MotoGP.com API or alternative data sources
"""

import asyncio
from datetime import datetime
from typing import List, Dict, Optional, Any
import aiohttp
from bs4 import BeautifulSoup

from src.config import settings
from src.utils.logger import logger


class MotoGPAPIClient:
    """
    Client for fetching MotoGP data
    
    Note: This is a template. The actual implementation will depend on:
    - Official API access (requires authentication)
    - Alternative APIs (unofficial)
    - Web scraping as fallback
    """
    
    def __init__(self):
        self.api_url = settings.motogp_api_url
        self.api_key = settings.motogp_api_key
        self.api_secret = settings.motogp_api_secret
        self.session: Optional[aiohttp.ClientSession] = None
    
    async def __aenter__(self):
        """Async context manager entry"""
        self.session = aiohttp.ClientSession()
        return self
    
    async def __aexit__(self, exc_type, exc_val, exc_tb):
        """Async context manager exit"""
        if self.session:
            await self.session.close()
    
    async def _make_request(
        self,
        endpoint: str,
        method: str = "GET",
        params: Optional[Dict] = None,
        data: Optional[Dict] = None
    ) -> Dict[str, Any]:
        """Make HTTP request to API"""
        if not self.session:
            raise RuntimeError("Client session not initialized. Use async context manager.")
        
        url = f"{self.api_url}/{endpoint}"
        headers = {}
        
        if self.api_key:
            headers["X-API-Key"] = self.api_key
        
        try:
            async with self.session.request(
                method,
                url,
                params=params,
                json=data,
                headers=headers
            ) as response:
                response.raise_for_status()
                return await response.json()
        except aiohttp.ClientError as e:
            logger.error(f"API request error: {e}")
            raise
    
    async def get_current_season(self) -> int:
        """Get current season year"""
        # TODO: Implement based on actual API
        return settings.current_season
    
    async def get_calendar(self, season: int) -> List[Dict[str, Any]]:
        """
        Get race calendar for a season
        
        Returns:
            List of events with circuit info and dates
        """
        # TODO: Implement based on actual API
        # This is a placeholder structure
        logger.info(f"Fetching calendar for season {season}")
        
        # Example return structure:
        return [
            {
                "event_id": "QAT",
                "name": "Qatar GP",
                "circuit": "Losail International Circuit",
                "country": "Qatar",
                "date": "2024-03-10",
                "races": {
                    "motogp": {
                        "sprint": "2024-03-09T15:00:00",
                        "race": "2024-03-10T14:00:00"
                    },
                    "moto2": {
                        "race": "2024-03-10T12:15:00"
                    },
                    "moto3": {
                        "race": "2024-03-10T11:00:00"
                    }
                }
            }
        ]
    
    async def get_riders(self, season: int, category: str) -> List[Dict[str, Any]]:
        """
        Get riders for a specific category and season
        
        Args:
            season: Year
            category: MOTOGP, MOTO2, or MOTO3
        
        Returns:
            List of riders with details
        """
        logger.info(f"Fetching riders for {category} season {season}")
        
        # TODO: Implement based on actual API
        # Example return structure:
        return [
            {
                "rider_id": "93",
                "number": 93,
                "first_name": "Marc",
                "last_name": "Márquez",
                "country": "Spain",
                "team": "Repsol Honda Team",
                "bike": "Honda RC213V"
            }
        ]
    
    async def get_session_results(
        self,
        event_id: str,
        category: str,
        session_type: str
    ) -> List[Dict[str, Any]]:
        """
        Get results for a practice/qualifying session
        
        Args:
            event_id: Event identifier
            category: MOTOGP, MOTO2, or MOTO3
            session_type: FP1, FP2, Q1, Q2, etc.
        
        Returns:
            List of session results with times and positions
        """
        logger.info(f"Fetching {session_type} results for {category} at {event_id}")
        
        # TODO: Implement based on actual API
        return [
            {
                "position": 1,
                "rider_id": "93",
                "rider_name": "M. Márquez",
                "best_lap": "1:38.838",
                "gap": "-",
                "laps": 24
            }
        ]
    
    async def get_race_results(
        self,
        event_id: str,
        category: str,
        race_type: str = "race"
    ) -> List[Dict[str, Any]]:
        """
        Get race results
        
        Args:
            event_id: Event identifier
            category: MOTOGP, MOTO2, or MOTO3
            race_type: 'sprint' or 'race'
        
        Returns:
            List of race results with positions
        """
        logger.info(f"Fetching {race_type} results for {category} at {event_id}")
        
        # TODO: Implement based on actual API
        return [
            {
                "position": 1,
                "rider_id": "93",
                "rider_name": "M. Márquez",
                "time": "41:08.985",
                "gap": "-",
                "points": 25,
                "status": "finished"
            }
        ]
    
    async def get_championship_standings(
        self,
        season: int,
        category: str
    ) -> List[Dict[str, Any]]:
        """
        Get championship standings
        
        Args:
            season: Year
            category: MOTOGP, MOTO2, or MOTO3
        
        Returns:
            List of riders with points
        """
        logger.info(f"Fetching championship standings for {category} {season}")
        
        # TODO: Implement based on actual API
        return [
            {
                "position": 1,
                "rider_id": "93",
                "rider_name": "M. Márquez",
                "points": 325,
                "team": "Repsol Honda Team"
            }
        ]


# Alternative: Web scraping fallback
class MotoGPScraperClient:
    """
    Fallback client using web scraping
    Use only if official API is not available
    """
    
    def __init__(self):
        self.base_url = "https://www.motogp.com"
        self.session: Optional[aiohttp.ClientSession] = None
    
    async def __aenter__(self):
        self.session = aiohttp.ClientSession()
        return self
    
    async def __aexit__(self, exc_type, exc_val, exc_tb):
        if self.session:
            await self.session.close()
    
    async def _fetch_page(self, url: str) -> str:
        """Fetch HTML page"""
        if not self.session:
            raise RuntimeError("Client session not initialized")
        
        try:
            async with self.session.get(url) as response:
                response.raise_for_status()
                return await response.text()
        except aiohttp.ClientError as e:
            logger.error(f"Scraping error: {e}")
            raise
    
    async def get_calendar_scrape(self, season: int) -> List[Dict[str, Any]]:
        """Scrape race calendar"""
        url = f"{self.base_url}/en/calendar"
        html = await self._fetch_page(url)
        
        # TODO: Implement actual scraping logic
        soup = BeautifulSoup(html, 'lxml')
        
        # Parse calendar data
        events = []
        
        return events


# Factory function to get appropriate client
def get_motogp_client() -> MotoGPAPIClient:
    """Get MotoGP API client instance"""
    if settings.motogp_api_key:
        logger.info("Using official MotoGP API client")
        return MotoGPAPIClient()
    else:
        logger.warning("No API key configured. Limited functionality available.")
        return MotoGPAPIClient()  # Will need manual data entry or scraping
